middlewares

+++ part4/blogs/controllers/blogs.js
@@ -5,26 +5,21 @@ const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
-        .find({}).populate('user')
+        .find({}).populate('user', { username: 1, name: 1, id: 1 })
 
     response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-    const body = request.body
+    const {title, author, url, likes} = request.body
 
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if(!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
-    }
-    
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes,
+        title,
+        author,
+        url,
+        likes,
         user: user._id
     })
 
@@ -36,8 +31,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+    if(blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        await User.findByIdAndUpdate(user._id, { $pull: { blogs: request.params.id } })
+        return response.status(204).end()
+    }
+    return response.status(401).json({ error: 'unauthorized' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
